refactor: Split math parser util file
